wip: cleanup 2
