+++ src/typegate/src/runtimes/typegraph.ts
@@ -6,17 +6,11 @@ import {
   type TypeGraphDS,
   type TypeMaterializer,
 } from "../typegraph/mod.ts";
-import { TypeKind } from "graphql";
-import { ensure } from "../utils.ts";
 import { Runtime } from "./Runtime.ts";
 import { ComputeStage } from "../engine/query_engine.ts";
 import {
   isEither,
   isEmptyObject,
-  isFunction,
-  isList,
-  isObject,
-  isOptional,
   isQuantifier,
   isScalar,
   isUnion,
@@ -32,45 +26,23 @@ import {
 } from "../typegraph/visitor.ts";
 import { distinctBy } from "@std/collections/distinct-by";
 import { isInjected } from "../typegraph/utils.ts";
-import type { InjectionNode, PolicyIndices } from "../typegraph/types.ts";
+import type { InjectionNode } from "../typegraph/types.ts";
+import { TypeFormatter } from "./typegraph/formatter.ts";
+import {
+  typeCustomScalar,
+  typeEmptyObjectScalar,
+} from "./typegraph/helpers.ts";
 
-type DeprecatedArg = { includeDeprecated?: boolean };
-
-const SCALAR_TYPE_MAP = {
-  boolean: "Boolean",
-  integer: "Int",
-  float: "Float",
-  string: "String",
-  file: "File",
-};
-
-function generateCustomScalar(type: TypeNode, idx: number) {
-  if (isScalar(type)) {
-    const id = type.type;
-    return {
-      title: `_${id[0].toUpperCase()}${id.slice(1)}`,
-      type: "object",
-      properties: {
-        [id]: idx,
-      },
-    } as ObjectNode;
-  }
-  throw `"${type.title}" of type "${type.type}" is not a scalar`;
-}
-
-type FieldInfo = {
-  name: string;
-  typeIdx: number;
-  policies: PolicyIndices[];
-};
+export type DeprecatedArg = { includeDeprecated?: boolean };
 
 export class TypeGraphRuntime extends Runtime {
   tg: TypeGraphDS;
-  private scalarIndex = new Map<string, number>();
+  formatter: TypeFormatter;
 
   private constructor(tg: TypeGraphDS) {
     super(TypeGraph.formatName(tg));
     this.tg = tg;
+    this.formatter = new TypeFormatter(tg);
   }
 
   static init(
@@ -88,34 +60,7 @@ export class TypeGraphRuntime extends Runtime {
     _waitlist: ComputeStage[],
     _verbose: boolean,
   ): ComputeStage[] {
-    const resolver: Resolver = (() => {
-      const name = stage.props.materializer?.name;
-      if (name === "getSchema") {
-        return this.getSchema;
-      }
-      if (name === "getType") {
-        return this.getType;
-      }
-
-      if (name === "resolver") {
-        return async ({ _: { parent } }) => {
-          const resolver = parent[stage.props.node];
-          const ret = typeof resolver === "function"
-            ? await resolver()
-            : resolver;
-          return ret;
-        };
-      }
-
-      return async ({ _: { parent } }) => {
-        const resolver = parent[stage.props.node];
-        const ret = typeof resolver === "function"
-          ? await resolver()
-          : resolver;
-        return ret;
-      };
-    })();
-
+    const resolver = this.#delegate(stage);
     return [
       new ComputeStage({
         ...stage.props,
@@ -124,7 +69,33 @@ export class TypeGraphRuntime extends Runtime {
     ];
   }
 
-  getSchema: Resolver = () => {
+  #delegate(stage: ComputeStage): Resolver {
+    const name = stage.props.materializer?.name;
+    switch (name) {
+      case "getSchema":
+        return this.#getSchemaResolver;
+      case "getType":
+        return this.#getTypeResolver;
+      case "resolver":
+        return async ({ _: { parent } }) => {
+          const resolver = parent[stage.props.node];
+          const ret = typeof resolver === "function"
+            ? await resolver()
+            : resolver;
+          return ret;
+        };
+      default:
+        return async ({ _: { parent } }) => {
+          const resolver = parent[stage.props.node];
+          const ret = typeof resolver === "function"
+            ? await resolver()
+            : resolver;
+          return ret;
+        };
+    }
+  }
+
+  #getSchemaResolver: Resolver = () => {
     const root = this.tg.types[0] as ObjectNode;
 
     const queries = this.tg.types[root.properties["query"]] as ObjectNode;
@@ -133,448 +104,152 @@ export class TypeGraphRuntime extends Runtime {
     return {
       // https://github.com/graphql/graphql-js/blob/main/src/type/introspection.ts#L36
       description: () => `${root.type} typegraph`,
-      types: () => {
-        // filter non-native GraphQL types
-        const filter = (
-          type: TypeNode,
-          input: {
-            injectionTree: Record<string, InjectionNode>;
-            path: string[];
-          } | null,
-        ) => {
-          return (
-            (input == null ||
-              !isInjected(this.tg, type, input.path, input.injectionTree)) &&
-            !isQuantifier(type)
-          );
-        };
-
-        const scalarTypeIndices = new Set<number>();
-        const inputTypeIndices = new Set<number>();
-        const regularTypeIndices = new Set<number>();
-
-        const inputRootTypeIndices = new Set<number>();
-        const outputRootTypeIndices = new Set<number>();
-
-        // decides whether or not custom object should be generated
-        let hasUnion = false;
-        let requireEmptyObject = false;
-
-        const myVisitor: TypeVisitorMap = {
-          [Type.FUNCTION]: ({ type }) => {
-            // TODO skip if policy check fails
-            // https://metatype.atlassian.net/browse/MET-119
-
-            // the struct input of a function never generates a GrahpQL type
-            // the actual inputs are the properties
-            inputRootTypeIndices.add(type.input);
-            outputRootTypeIndices.add(type.output);
-            visitTypes(
-              this.tg,
-              getChildTypes(this.tg.types[type.input]),
-              ({ type, idx }) => {
-                hasUnion ||= isUnion(type) || isEither(type);
-                if (isScalar(type)) {
-                  scalarTypeIndices.add(idx);
-                  this.scalarIndex.set(type.type, idx);
-                  return false;
-                }
-                // FIXME
-                if (filter(type, null)) {
-                  inputTypeIndices.add(idx);
-                }
-                return true;
-              },
-            );
-            return true;
-          },
-          default: ({ type, idx }) => {
-            requireEmptyObject ||= isEmptyObject(type);
-
-            if (
-              inputRootTypeIndices.has(idx) &&
-              !outputRootTypeIndices.has(idx)
-            ) {
-              return false;
-            }
-            // type is either a regular type or an input type reused in the output
-            hasUnion ||= isUnion(type) || isEither(type);
-            if (isScalar(type)) {
-              scalarTypeIndices.add(idx);
-              this.scalarIndex.set(type.type, idx);
-              return false;
-            }
-            // FIXME
-            if (filter(type, null)) {
-              regularTypeIndices.add(idx);
-            }
-            return true;
-          },
-        };
-
-        visitTypes(this.tg, getChildTypes(this.tg.types[0]), myVisitor);
-        const distinctScalars = distinctBy(
-          [...scalarTypeIndices].map((idx) => this.tg.types[idx]),
-          (t) => t.type, // for scalars: one GraphQL type per `type` not `title`
-        );
-        const scalarTypes = distinctScalars.map((type) =>
-          this.formatType(type, false, false)
-        );
-
-        const adhocCustomScalarTypes = hasUnion
-          ? distinctScalars.map((node) => {
-            const idx = this.scalarIndex.get(node.type)!;
-            const asObject = generateCustomScalar(node, idx);
-            return this.formatType(asObject, false, false);
-          })
-          : [];
-
-        const regularTypes = distinctBy(
-          [...regularTypeIndices].map((idx) => this.tg.types[idx]),
-          (t) => t.title,
-        ).map((type) => this.formatType(type, false, false));
-
-        const inputTypes = distinctBy(
-          [...inputTypeIndices].map((idx) => this.tg.types[idx]),
-          (t) => t.title,
-        ).map((type) => this.formatType(type, false, true));
-
-        const types = [
-          ...scalarTypes,
-          ...adhocCustomScalarTypes,
-          ...regularTypes,
-          ...inputTypes,
-        ];
-
-        // Handle non-root leaf case
-        if (
-          requireEmptyObject &&
-          !types.some((t: any) => t!.title == "EmptyObject")
-        ) {
-          types.push(emptyObjectScalar());
-        }
-
-        this.scalarIndex.clear();
-        return types;
-      },
+      types: this.#typesResolver,
       queryType: () => {
         if (!queries || Object.values(queries.properties).length === 0) {
           // https://github.com/graphql/graphiql/issues/2308 (3x) enforce to keep empty Query type
-          return this.formatType(queries, false, false);
+          return this.formatter.formatType(queries, false, false);
         }
-        return this.formatType(queries, false, false);
+        return this.formatter.formatType(queries, false, false);
       },
       mutationType: () => {
         if (!mutations || Object.values(mutations.properties).length === 0) {
           return null;
         }
-        return this.formatType(mutations, false, false);
+        return this.formatter.formatType(mutations, false, false);
       },
       subscriptionType: () => null,
       directives: () => [],
     };
   };
 
-  getType: Resolver = ({ name }) => {
-    const type = this.tg.types.find((type) => type.title === name);
-    return type ? this.formatType(type, false, false) : null;
-  };
-
-  formatInputFields = (
-    [name, typeIdx]: [string, number],
-    injectionNode: InjectionNode | null,
-  ) => {
-    const type = this.tg.types[typeIdx];
-
-    // TODO resolve quantifiers
-
-    if (injectionNode && ("injection" in injectionNode)) {
-      // FIXME MET-704
-
-      return null;
-    }
-
-    return {
-      // https://github.com/graphql/graphql-js/blob/main/src/type/introspection.ts#L374
-      name: () => name,
-      description: () => `${name} input field`,
-      type: () => {
-        const ret = this.formatType(type, true, true);
-        return ret;
-      },
-      defaultValue: () => null,
-      isDeprecated: () => false,
-      deprecationReason: () => null,
-    };
-  };
-
-  formatType = (
-    type: TypeNode,
-    required: boolean,
-    asInput: boolean,
-  ): Record<string, () => unknown> => {
-    const common = {
-      // https://github.com/graphql/graphql-js/blob/main/src/type/introspection.ts#L207
-      name: () => null,
-      specifiedByURL: () => null,
-      // logic at https://github.com/graphql/graphql-js/blob/main/src/type/introspection.ts#L453-L490
-      ofType: () => null,
-      inputFields: () => null,
-      fields: () => null,
-      interfaces: () => null,
-      possibleTypes: () => null,
-      enumValues: () => null,
+  #typesResolver: Resolver = (_) => {
+    // filter non-native GraphQL types
+    const filter = (
+      type: TypeNode,
+      input: {
+        injectionTree: Record<string, InjectionNode>;
+        path: string[];
+      } | null,
+    ) => {
+      return (
+        (input == null ||
+          !isInjected(this.tg, type, input.path, input.injectionTree)) &&
+        !isQuantifier(type)
+      );
     };
 
-    if (isOptional(type)) {
-      const subtype = this.tg.types[type.item];
-      return this.formatType(subtype, false, asInput);
-    }
+    const scalarTypeIndices = new Set<number>();
+    const inputTypeIndices = new Set<number>();
+    const regularTypeIndices = new Set<number>();
 
-    if (required) {
-      return {
-        ...common,
-        kind: () => TypeKind.NON_NULL,
-        ofType: () => {
-          return this.formatType(type, false, asInput);
-        },
-      };
-    }
+    const inputRootTypeIndices = new Set<number>();
+    const outputRootTypeIndices = new Set<number>();
 
-    if (isList(type)) {
-      return {
-        ...common,
-        kind: () => TypeKind.LIST,
-        ofType: () => {
-          const subtype = this.tg.types[type.items];
-          return this.formatType(subtype, true, asInput);
-        },
-      };
-    }
+    // decides whether or not custom object should be generated
+    let hasUnion = false;
+    let requireEmptyObject = false;
 
-    if (isScalar(type)) {
-      return {
-        ...common,
-        kind: () => TypeKind.SCALAR,
-        name: () => SCALAR_TYPE_MAP[type.type],
-        description: () => `${type.type} type`,
-      };
-    }
+    const myVisitor: TypeVisitorMap = {
+      [Type.FUNCTION]: ({ type }) => {
+        // TODO skip if policy check fails
+        // https://metatype.atlassian.net/browse/MET-119
 
-    if (isFunction(type)) {
-      const output = this.tg.types[type.output as number];
-      return this.formatType(output, false, false);
-    }
-
-    if (isObject(type)) {
-      if (type.title === "Query" && Object.keys(type.properties).length === 0) {
-        // https://github.com/graphql/graphiql/issues/2308 (3x) enforce to keep empty Query type
-        return {
-          ...common,
-          kind: () => TypeKind.OBJECT,
-          name: () => "Query",
-          description: () => `${type.title} type`,
-          fields: () => [
-            {
-              name: () => "_",
-              args: () => [],
-              type: () =>
-                this.formatType(
-                  this.tg.types[
-                    (this.tg.types[0] as ObjectNode).properties["query"]
-                  ], // itself
-                  false,
-                  false,
-                ),
-              isDeprecated: () => true,
-              deprecationReason: () =>
-                "Dummy value due to https://github.com/graphql/graphiql/issues/2308",
-            },
-          ],
-          interfaces: () => [],
-        };
-      }
-
-      if (isEmptyObject(type)) {
-        return emptyObjectScalar();
-      }
-
-      if (asInput) {
-        return {
-          ...common,
-          kind: () => TypeKind.INPUT_OBJECT,
-          name: () => `${type.title}Inp`,
-          description: () => `${type.title} input type`,
-          inputFields: () => {
-            return Object.entries(type.properties).map(([name, typeIdx]) =>
-              this.formatField(true)({
-                name,
-                typeIdx,
-                policies: type.policies?.[name] ?? [],
-              })
-            );
+        // the struct input of a function never generates a GrahpQL type
+        // the actual inputs are the properties
+        inputRootTypeIndices.add(type.input);
+        outputRootTypeIndices.add(type.output);
+        visitTypes(
+          this.tg,
+          getChildTypes(this.tg.types[type.input]),
+          ({ type, idx }) => {
+            hasUnion ||= isUnion(type) || isEither(type);
+            if (isScalar(type)) {
+              scalarTypeIndices.add(idx);
+              this.formatter.scalarIndex.set(type.type, idx);
+              return false;
+            }
+            // FIXME
+            if (filter(type, null)) {
+              inputTypeIndices.add(idx);
+            }
+            return true;
           },
-          interfaces: () => [],
-        };
-      } else {
-        return {
-          ...common,
-          kind: () => TypeKind.OBJECT,
-          name: () => type.title,
-          description: () => `${type.title} type`,
-          fields: () => {
-            let entries = Object.entries(type.properties);
-            entries = entries.sort((a, b) => b[1] - a[1]);
-            return entries.map(([name, typeIdx]) =>
-              this.formatField(false)({
-                name,
-                typeIdx,
-                policies: type.policies?.[name] ?? [],
-              })
-            );
-          },
-          interfaces: () => [],
-        };
-      }
-    }
-
-    // Issue:
-    // Translate union (anyOf) / either (oneOf) to Graphql types that behave the same way
-    // - Current graphql spec does not allow UNION types on the input (yet)
-    // - UNION types does not support scalars, only objects
-    // Current solution:
-    // - input: translate either/union nodes to a custom scalar
-    // - output: translate either/union nodes to graphql UNION
-    //  * caveat: since UNION does not allow scalar variants, we have to translate them
-    //    to custom graphql objects
-    // Alternative ?
-    // https://github.com/graphql/graphql-spec/pull/825
-    if (isEither(type) || isUnion(type)) {
-      const variants = isUnion(type) ? type.anyOf : type.oneOf;
-      if (asInput) {
-        const titles = new Set<string>(
-          variants.map((idx) => this.tg.types[idx].title),
         );
-        const description = `${type.type} type\n${
-          Array.from(titles).join(
-            ", ",
-          )
-        }`;
+        return true;
+      },
+      default: ({ type, idx }) => {
+        requireEmptyObject ||= isEmptyObject(type);
 
-        return {
-          ...common,
-          kind: () => TypeKind.SCALAR,
-          name: () => `${type.title}In`,
-          description: () => description,
-        };
-      } else {
-        return {
-          ...common,
-          kind: () => TypeKind.UNION,
-          name: () => `${type.title}Out`,
-          description: () => `${type.title} type`,
-          possibleTypes: () => {
-            return variants.map((idx) => {
-              const variant = this.tg.types[idx];
-              if (isScalar(variant)) {
-                const idx = this.scalarIndex.get(variant.type)!;
-                const asObject = generateCustomScalar(variant, idx);
-                return this.formatType(asObject, false, false);
-              } else {
-                return this.formatType(variant, false, false);
-              }
-            });
-          },
-        };
-      }
+        if (
+          inputRootTypeIndices.has(idx) &&
+          !outputRootTypeIndices.has(idx)
+        ) {
+          return false;
+        }
+        // type is either a regular type or an input type reused in the output
+        hasUnion ||= isUnion(type) || isEither(type);
+        if (isScalar(type)) {
+          scalarTypeIndices.add(idx);
+          this.formatter.scalarIndex.set(type.type, idx);
+          return false;
+        }
+        // FIXME
+        if (filter(type, null)) {
+          regularTypeIndices.add(idx);
+        }
+        return true;
+      },
+    };
+
+    visitTypes(this.tg, getChildTypes(this.tg.types[0]), myVisitor);
+    const distinctScalars = distinctBy(
+      [...scalarTypeIndices].map((idx) => this.tg.types[idx]),
+      (t) => t.type, // for scalars: one GraphQL type per `type` not `title`
+    );
+    const scalarTypes = distinctScalars.map((type) =>
+      this.formatter.formatType(type, false, false)
+    );
+
+    const adhocCustomScalarTypes = hasUnion
+      ? distinctScalars.map((node) => {
+        const idx = this.formatter.scalarIndex.get(node.type)!;
+        const asObject = typeCustomScalar(node, idx);
+        return this.formatter.formatType(asObject, false, false);
+      })
+      : [];
+
+    const regularTypes = distinctBy(
+      [...regularTypeIndices].map((idx) => this.tg.types[idx]),
+      (t) => t.title,
+    ).map((type) => this.formatter.formatType(type, false, false));
+
+    const inputTypes = distinctBy(
+      [...inputTypeIndices].map((idx) => this.tg.types[idx]),
+      (t) => t.title,
+    ).map((type) => this.formatter.formatType(type, false, true));
+
+    const types = [
+      ...scalarTypes,
+      ...adhocCustomScalarTypes,
+      ...regularTypes,
+      ...inputTypes,
+    ];
+
+    // Handle non-root leaf case
+    if (
+      requireEmptyObject &&
+      !types.some((t: any) => t!.title == "EmptyObject")
+    ) {
+      types.push(typeEmptyObjectScalar());
     }
 
-    throw Error(`unexpected type format ${(type as any).type}`);
-    // interface: fields, interfaces, possibleTypes
-    // union: possibleTypes
-    // enum: enumValues
+    this.formatter.scalarIndex.clear();
+    return types;
   };
 
-  policyDescription(policies: PolicyIndices[]): string {
-    const describeOne = (p: number) => this.tg.policies[p].name;
-    const describe = (p: PolicyIndices) => {
-      if (typeof p === "number") {
-        return describeOne(p);
-      }
-      return Object.entries(p)
-        .map(([eff, polIdx]) => `${eff}:${describeOne(polIdx)}`)
-        .join("; ");
-    };
-    const policyNames = policies.map(describe);
-
-    let ret = "\n\nPolicies:\n";
-
-    if (policyNames.length > 0) {
-      ret += policyNames.map((p: string) => `- ${p}`).join("\n");
-    } else {
-      ret += "- inherit";
-    }
-
-    return ret;
-  }
-
-  formatField =
-    (asInput: boolean) => ({ name, typeIdx, policies }: FieldInfo) => {
-      const type = this.tg.types[typeIdx];
-      const common = {
-        // https://github.com/graphql/graphql-js/blob/main/src/type/introspection.ts#L329
-        name: () => name,
-        description: () => `${name} field${this.policyDescription(policies)}`,
-        isDeprecated: () => false,
-        deprecationReason: () => null,
-      };
-
-      if (isFunction(type)) {
-        return {
-          ...common,
-          args: (_: DeprecatedArg = {}) => {
-            const inp = this.tg.types[type.input as number];
-            ensure(
-              isObject(inp),
-              `${type} cannot be an input field, require struct`,
-            );
-            let entries = Object.entries((inp as ObjectNode).properties);
-            entries = entries.sort((a, b) => b[1] - a[1]);
-            return entries
-              .map((entry) =>
-                this.formatInputFields(
-                  entry,
-                  (type.injections ?? {})[entry[0]] ?? null,
-                )
-              )
-              .filter((f) => f !== null);
-          },
-          type: () => {
-            const output = this.tg.types[type.output as number];
-            return this.formatType(output, true, false);
-          },
-        };
-      }
-
-      return {
-        ...common,
-        args: () => [],
-        type: () => {
-          return this.formatType(type, true, asInput);
-        },
-      };
-    };
-}
-
-function emptyObjectScalar() {
-  return {
-    kind: () => TypeKind.SCALAR,
-    name: () => "EmptyObject",
-    description: () => "object scalar type representing an empty object",
-    fields: () => {},
-    inputFields: () => {},
-    interfaces: () => {},
-    enumValues: () => {},
-    possibleTypes: () => {},
+  #getTypeResolver: Resolver = ({ name }) => {
+    const type = this.tg.types.find((type) => type.title === name);
+    return type ? this.formatter.formatType(type, false, false) : null;
   };
 }
